query/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0